requirements.txt
# Discord
discord.py==2.3.2
py-cord==2.5.0

# Exchange APIs
python-binance==1.0.19
pybit==5.6.2
gate-api==4.41.3

# Technical Analysis
ta==0.11.0
ta-lib==0.4.28
pandas-ta==0.3.14b0
pandas==2.1.4
numpy==1.26.2

# Utilities
python-dotenv==1.0.0
colorlog==6.8.0
aiohttp==3.9.1
asyncio==3.4.3

# Database
aiosqlite==0.19.0

# Plotting
plotly==5.18.0
matplotlib==3.8.2

# Tables
tabulate==0.9.0
